Fjernet gamle filer for feltformatererPdfKvittering, pdfKvitteringService og søknadTreeWalker.
