+++ src/main/kotlin/no/nav/familie/ef/mottak/service/PdfService.kt
@@ -5,6 +5,7 @@ import no.nav.familie.ef.mottak.config.DOKUMENTTYPE_OVERGANGSSTØNAD
 import no.nav.familie.ef.mottak.config.DOKUMENTTYPE_SKJEMA_ARBEIDSSØKER
 import no.nav.familie.ef.mottak.config.DOKUMENTTYPE_SKOLEPENGER
 import no.nav.familie.ef.mottak.integration.PdfClient
+import no.nav.familie.ef.mottak.integration.PdfKvitteringClient
 import no.nav.familie.ef.mottak.mapper.SøknadMapper
 import no.nav.familie.ef.mottak.repository.EttersendingRepository
 import no.nav.familie.ef.mottak.repository.SøknadRepository
@@ -26,14 +27,24 @@ class PdfService(
     private val ettersendingRepository: EttersendingRepository,
     private val vedleggRepository: VedleggRepository,
     private val pdfClient: PdfClient,
+    private val pdfKvitteringClient: PdfKvitteringClient,
 ) {
     fun lagPdf(id: String) {
         val innsending = søknadRepository.findByIdOrNull(id) ?: error("Kunne ikke finne søknad ($id) i database")
         val vedleggTitler = vedleggRepository.finnTitlerForSøknadId(id).sorted()
         val feltMap = lagFeltMap(innsending, vedleggTitler)
-        val søknadPdf = pdfClient.lagPdf(feltMap)
+        val søknadPdf = pdfKvitteringClient.opprettPdf(feltMap)
         val oppdatertSoknad = innsending.copy(søknadPdf = EncryptedFile(søknadPdf))
         søknadRepository.update(oppdatertSoknad)
+    }
+
+    fun lagForsideForEttersending(
+        ettersending: Ettersending,
+        vedleggTitler: List<String>,
+    ) {
+        val feltMap = SøknadTilFeltMap.mapEttersending(ettersending, vedleggTitler)
+        val søknadPdf = pdfClient.lagPdf(feltMap)
+        ettersendingRepository.update(ettersending.copy(ettersendingPdf = EncryptedFile(søknadPdf)))
     }
 
     private fun lagFeltMap(
@@ -43,31 +54,22 @@ class PdfService(
         when (innsending.dokumenttype) {
             DOKUMENTTYPE_OVERGANGSSTØNAD -> {
                 val dto = SøknadMapper.toDto<SøknadOvergangsstønad>(innsending)
-                SøknadTreeWalker.mapOvergangsstønad(dto, vedleggTitler)
+                SøknadTilFeltMap.mapOvergangsstønad(dto, vedleggTitler)
             }
             DOKUMENTTYPE_BARNETILSYN -> {
                 val dto = SøknadMapper.toDto<SøknadBarnetilsyn>(innsending)
-                SøknadTreeWalker.mapBarnetilsyn(dto, vedleggTitler)
+                SøknadTilFeltMap.mapBarnetilsyn(dto, vedleggTitler)
             }
             DOKUMENTTYPE_SKJEMA_ARBEIDSSØKER -> {
                 val dto = SøknadMapper.toDto<SkjemaForArbeidssøker>(innsending)
-                SøknadTreeWalker.mapSkjemafelter(dto)
+                SøknadTilFeltMap.mapSkjemafelter(dto)
             }
             DOKUMENTTYPE_SKOLEPENGER -> {
                 val dto = SøknadMapper.toDto<SøknadSkolepenger>(innsending)
-                SøknadTreeWalker.mapSkolepenger(dto, vedleggTitler)
+                SøknadTilFeltMap.mapSkolepenger(dto, vedleggTitler)
             }
             else -> {
                 error("Ukjent eller manglende dokumenttype id: ${innsending.id}")
             }
         }
-
-    fun lagForsideForEttersending(
-        ettersending: Ettersending,
-        vedleggTitler: List<String>,
-    ) {
-        val feltMap = SøknadTreeWalker.mapEttersending(ettersending, vedleggTitler)
-        val søknadPdf = pdfClient.lagPdf(feltMap)
-        ettersendingRepository.update(ettersending.copy(ettersendingPdf = EncryptedFile(søknadPdf)))
-    }
 }
