+++ src/main/kotlin/no/nav/familie/ef/mottak/task/LagSøknadPdfTask.kt
@@ -1,6 +1,6 @@
 package no.nav.familie.ef.mottak.task
 
-import no.nav.familie.ef.mottak.service.PdfKvitteringService
+import no.nav.familie.ef.mottak.service.PdfService
 import no.nav.familie.prosessering.AsyncTaskStep
 import no.nav.familie.prosessering.TaskStepBeskrivelse
 import no.nav.familie.prosessering.domene.Task
@@ -8,13 +8,13 @@ import no.nav.familie.prosessering.internal.TaskService
 import org.springframework.stereotype.Service
 
 @Service
-@TaskStepBeskrivelse(taskStepType = LagSøknadPdfTask.TYPE, beskrivelse = "Lag pdf av søknad")
+@TaskStepBeskrivelse(taskStepType = LagSøknadPdfTask.TYPE, beskrivelse = "Lag pdf-oppsummering av søknad")
 class LagSøknadPdfTask(
-    private val pdfKvitteringService: PdfKvitteringService,
+    private val pdfService: PdfService,
     private val taskService: TaskService,
 ) : AsyncTaskStep {
     override fun doTask(task: Task) {
-        pdfKvitteringService.lagPdf(task.payload)
+        pdfService.lagPdf(task.payload)
     }
 
     override fun onCompletion(task: Task) {
